Allow user-provided IHostBuilder
In order to integrate custom DI container such as Autofac, one needs to
be able to customize creation of the `IServiceProvider` instance.

A new constructor overload for `CommandLineApplication` accepts an
instance of `IHostBulder`, which can be used to execute the
customization necessary to integrate other DI libraries. For example,
one may invoke `UseServiceProviderFactory()` directly on the host
builder for this.

See the example below which integrates `Autofac`:

```cs
var hostBuilder = Host.CreateDefaultBuilder()
    .UseServiceProviderFactory(new AutofacServiceProviderFactory(CompositionRoot.Setup));

return new CommandLineApplication(hostBuilder)
    .AddCommand<VersionCommandHandler, VersionCommand>()
    .AddCommandGroup("list", "List information from the guide",
        cli =>
        {
            cli.AddCommand<ListCustomFormatsCommandHandler, ListCustomFormatsCommand>();
        });
```

Where `CompositionRoot.Setup()` sets up registrations in the Autofac
container.